fix: fix circular imports
- refactor code to eliminate circular imports
- add tool to check for circular imports and also run as part of CI/CD code quality checks
