+++ .github/workflows/check-code-quality.yml
@@ -35,4 +35,6 @@ jobs:
 
       - run: yarn typecheck
 
+      - run: yarn circular-import-check
+
       - run: yarn test:unit
