+++ package.json
@@ -73,6 +73,7 @@
     "babel-jest": "^29.7.0",
     "cross-env": "^7.0.3",
     "dotenv": "^16.3.1",
+    "dpdm": "^3.14.0",
     "electron": "^27.1.0",
     "electron-builder": "^24.9.0",
     "electron-playwright-helpers": "^1.7.0",
@@ -106,6 +107,7 @@
   "scripts": {
     "start": "yarn generate-notice && tsc -p src/ElectronBackend && vite",
     "build": "yarn generate-notice && vite build && tsc -p src/ElectronBackend",
+    "circular-import-check": "dpdm src/**/*.ts[x] --circular -T --tree false --warning false --exit-code circular:1",
     "typecheck": "tsc -p ./ && tsc --noEmit -p src/ElectronBackend",
     "test:unit": "jest",
     "test:changed": "jest --onlyChanged",
