+++ src/ElectronBackend/errorHandling/__tests__/errorHandling.test.ts
@@ -3,20 +3,15 @@
 // SPDX-FileCopyrightText: Nico Carl <nicocarl@protonmail.com>
 //
 // SPDX-License-Identifier: Apache-2.0
-import { app, BrowserWindow, dialog, WebContents } from 'electron';
+import { BrowserWindow, dialog, WebContents } from 'electron';
 
 import { AllowedFrontendChannels } from '../../../shared/ipc-channels';
 import { SendErrorInformationArgs } from '../../../shared/shared-types';
 import { loadInputAndOutputFromFilePath } from '../../input/importFromFile';
-import { getOpenFileListener } from '../../main/listeners';
-import { JsonParsingError } from '../../types/types';
 import {
   createListenerCallbackWithErrorHandling,
   getMessageBoxContentForErrorsWrapper,
-  getMessageBoxContentForParsingError,
   getMessageBoxForErrors,
-  getMessageBoxForInvalidDotOpossumFileError,
-  getMessageBoxForParsingError,
 } from '../errorHandling';
 
 jest.mock('electron', () => ({
@@ -149,82 +144,6 @@ describe('error handling', () => {
         AllowedFrontendChannels.RestoreFrontend,
       );
       expect(loadInputAndOutputFromFilePath).toBeCalled();
-    });
-  });
-
-  describe('getMessageBoxContentForParsingErrors', () => {
-    it('display an error message', () => {
-      const testMessage = 'Test error message';
-      const messageBoxContentForParsingErrors =
-        getMessageBoxContentForParsingError(testMessage);
-      expect(messageBoxContentForParsingErrors.detail).toContain(
-        'Test error message',
-      );
-      expect(messageBoxContentForParsingErrors.message).toEqual(
-        'Error parsing the input file.',
-      );
-    });
-  });
-
-  describe('getMessageBoxForParsingError', () => {
-    it('returns a messageBox', async () => {
-      const parsingError: JsonParsingError = {
-        message: 'parsingErrorMessage',
-        type: 'jsonParsingError',
-      };
-
-      await getMessageBoxForParsingError(parsingError.message);
-
-      expect(dialog.showMessageBox).toBeCalledWith(
-        expect.objectContaining({
-          type: 'error',
-          message: 'Error parsing the input file.',
-          detail: 'parsingErrorMessage',
-          buttons: ['Ok'],
-        }),
-      );
-      expect(app.exit).toBeCalledWith(0);
-    });
-  });
-
-  describe('getMessageBoxForInvalidDotOpossumFileError', () => {
-    it('returns a message box with correct content', async () => {
-      const testFilesInArchive = 'inpt.json, output.json';
-      const mainWindow = {
-        webContents: { send: jest.fn() },
-      } as unknown as BrowserWindow;
-
-      await getMessageBoxForInvalidDotOpossumFileError(
-        testFilesInArchive,
-        mainWindow,
-      );
-
-      expect(dialog.showMessageBox).toBeCalledWith(
-        expect.objectContaining({
-          type: 'error',
-          message: "Error loading '.opossum' file.",
-          detail:
-            "The '.opossum' file is invalid as it does not contain an 'input.json'. " +
-            `Actual files in the archive: ${testFilesInArchive}. ` +
-            'Either open another file or quit the application.',
-          buttons: ['Open New File', 'Quit'],
-        }),
-      );
-
-      expect(getOpenFileListener).toBeCalledWith(mainWindow);
-    });
-
-    it("invokes getOpenFileListener with mainWindow if 'Load New File' is clicked", async () => {
-      const testFilesInArchive = 'inpt.json, output.json';
-      const mainWindow = {
-        webContents: { send: jest.fn() },
-      } as unknown as BrowserWindow;
-
-      await getMessageBoxForInvalidDotOpossumFileError(
-        testFilesInArchive,
-        mainWindow,
-      );
-      expect(getOpenFileListener).toBeCalledWith(mainWindow);
     });
   });
 });
