+++ src/ElectronBackend/errorHandling/errorHandling.ts
@@ -15,7 +15,6 @@ import log from 'electron-log';
 import { AllowedFrontendChannels } from '../../shared/ipc-channels';
 import { loadInputAndOutputFromFilePath } from '../input/importFromFile';
 import { getGlobalBackendState } from '../main/globalBackendState';
-import { getOpenFileListener } from '../main/listeners';
 import { getLoadedFilePath } from '../utils/getLoadedFile';
 
 export function createListenerCallbackWithErrorHandling(
@@ -103,79 +102,6 @@ function performButtonAction(
         mainWindow,
         getLoadedFilePath(globalBackendState) as string,
       );
-      break;
-    case 1:
-      app.exit(0);
-      break;
-    default:
-      return;
-  }
-}
-
-export function getMessageBoxForParsingError(
-  errorMessage: string,
-): Promise<void> {
-  return getErrorDialog(
-    getMessageBoxContentForParsingError,
-    errorMessage,
-    () => {
-      app.exit(0);
-    },
-  );
-}
-
-export function getMessageBoxContentForParsingError(
-  errorMessage: string,
-): MessageBoxOptions {
-  return {
-    type: 'error',
-    buttons: ['Ok'],
-    defaultId: 0,
-    title: 'Parsing Error',
-    message: 'Error parsing the input file.',
-    detail: `${errorMessage}`,
-  };
-}
-
-export async function getMessageBoxForInvalidDotOpossumFileError(
-  filesInArchive: string,
-  mainWindow: BrowserWindow,
-): Promise<void> {
-  log.info("Error loading '.opossum' file. No 'input.json' found.");
-  return getErrorDialog(
-    getMessageBoxContentForInvalidDotOpossumFileError,
-    filesInArchive,
-    (value: MessageBoxReturnValue) =>
-      performButtonActionForInvalidDotOpossumFileError(
-        mainWindow,
-        value.response,
-      ),
-  );
-}
-
-function getMessageBoxContentForInvalidDotOpossumFileError(
-  filesInArchive: string,
-): MessageBoxOptions {
-  return {
-    type: 'error',
-    buttons: ['Open New File', 'Quit'],
-    defaultId: 0,
-    title: 'Invalid File Error',
-    message: "Error loading '.opossum' file.",
-    detail:
-      "The '.opossum' file is invalid as it does not contain an 'input.json'. " +
-      `Actual files in the archive: ${filesInArchive}. ` +
-      'Either open another file or quit the application.',
-  };
-}
-
-async function performButtonActionForInvalidDotOpossumFileError(
-  mainWindow: BrowserWindow,
-  buttonIndex: number,
-): Promise<void> {
-  switch (buttonIndex) {
-    case 0:
-      await getOpenFileListener(mainWindow)();
       break;
     case 1:
       app.exit(0);
