+++ src/ElectronBackend/input/__tests__/importFromFile.test.ts
@@ -19,7 +19,6 @@ import {
   PackageInfo,
   ParsedFileContent,
 } from '../../../shared/shared-types';
-import { getMessageBoxForParsingError } from '../../errorHandling/errorHandling';
 import {
   getGlobalBackendState,
   setGlobalBackendState,
@@ -30,8 +29,16 @@ import {
   deleteFolder,
   writeOpossumFile,
 } from '../../test-helpers';
-import { OpossumOutputFile, ParsedOpossumInputFile } from '../../types/types';
-import { loadInputAndOutputFromFilePath } from '../importFromFile';
+import {
+  JsonParsingError,
+  OpossumOutputFile,
+  ParsedOpossumInputFile,
+} from '../../types/types';
+import {
+  getMessageBoxForInvalidDotOpossumFileError,
+  getMessageBoxForParsingError,
+  loadInputAndOutputFromFilePath,
+} from '../importFromFile';
 
 const externalAttributionUuid = 'ecd692d9-b154-4d4d-be8c-external';
 const manualAttributionUuid = 'ecd692d9-b154-4d4d-be8c-manual';
@@ -44,10 +51,7 @@ jest.mock('electron', () => ({
   BrowserWindow: {
     getFocusedWindow: jest.fn(),
   },
-  app: {
-    getName: jest.fn(),
-    getVersion: jest.fn(),
-  },
+  app: { exit: jest.fn(), getName: jest.fn(), getVersion: jest.fn() },
 }));
 
 jest.mock('electron-log');
@@ -218,7 +222,6 @@ describe('Test of loading function', () => {
       expectedNumberOfCalls,
     );
 
-    expect(getMessageBoxForParsingError).toHaveBeenCalled();
     expect(getGlobalBackendState()).toEqual(expectedBackendState);
     deleteFolder(temporaryPath);
   });
@@ -637,3 +640,42 @@ function assertFileLoadedCorrectly(testUuid: string): void {
   );
   expect(dialog.showMessageBox).not.toBeCalled();
 }
+
+describe('getMessageBoxForParsingError', () => {
+  it('returns a messageBox', async () => {
+    const parsingError: JsonParsingError = {
+      message: 'parsingErrorMessage',
+      type: 'jsonParsingError',
+    };
+
+    await getMessageBoxForParsingError(parsingError.message);
+
+    expect(dialog.showMessageBox).toBeCalledWith(
+      expect.objectContaining({
+        type: 'error',
+        message: 'Error parsing the input file.',
+        detail: 'parsingErrorMessage',
+        buttons: ['OK'],
+      }),
+    );
+  });
+});
+
+describe('getMessageBoxForInvalidDotOpossumFileError', () => {
+  it('returns a message box with correct content', async () => {
+    const testFilesInArchive = 'inpt.json, output.json';
+
+    await getMessageBoxForInvalidDotOpossumFileError(testFilesInArchive);
+
+    expect(dialog.showMessageBox).toBeCalledWith(
+      expect.objectContaining({
+        type: 'error',
+        message: "Error loading '.opossum' file.",
+        detail:
+          "The '.opossum' file is invalid as it does not contain an 'input.json'. " +
+          `Actual files in the archive: ${testFilesInArchive}.`,
+        buttons: ['OK'],
+      }),
+    );
+  });
+});
