+++ src/Frontend/Components/AttributionColumn/AttributionColumn.tsx
@@ -24,11 +24,9 @@ import {
   getTemporaryDisplayPackageInfo,
   wereTemporaryDisplayPackageInfoModified,
 } from '../../state/selectors/all-views-resource-selectors';
-import { getSelectedAttributionIdInAttributionView } from '../../state/selectors/attribution-view-resource-selectors';
-import {
-  getDisplayedPackage,
-  getResolvedExternalAttributions,
-} from '../../state/selectors/audit-view-resource-selectors';
+import { getSelectedAttributionIdInAttributionView } from '../../state/selectors/all-views-resource-selectors';
+import { getDisplayedPackage } from '../../state/selectors/all-views-resource-selectors';
+import { getResolvedExternalAttributions } from '../../state/selectors/audit-view-resource-selectors';
 import {
   getQAMode,
   getSelectedView,
