+++ src/Frontend/Components/AttributionDetailsViewer/AttributionDetailsViewer.tsx
@@ -16,10 +16,8 @@ import {
 import { openPopup } from '../../state/actions/view-actions/view-actions';
 import { useAppDispatch, useAppSelector } from '../../state/hooks';
 import { getTemporaryDisplayPackageInfo } from '../../state/selectors/all-views-resource-selectors';
-import {
-  getResourceIdsOfSelectedAttribution,
-  getSelectedAttributionIdInAttributionView,
-} from '../../state/selectors/attribution-view-resource-selectors';
+import { getSelectedAttributionIdInAttributionView } from '../../state/selectors/all-views-resource-selectors';
+import { getResourceIdsOfSelectedAttribution } from '../../state/selectors/attribution-view-resource-selectors';
 import { convertDisplayPackageInfoToPackageInfo } from '../../util/convert-package-info';
 import { AttributionColumn } from '../AttributionColumn/AttributionColumn';
 import { ResizableBox } from '../ResizableBox/ResizableBox';
