+++ src/Frontend/Components/ConfirmDeletionPopup/ConfirmDeletionPopup.tsx
@@ -10,11 +10,9 @@ import {
   deleteAttributionGloballyAndSave,
 } from '../../state/actions/resource-actions/save-actions';
 import { useAppDispatch, useAppSelector } from '../../state/hooks';
-import { getSelectedAttributionIdInAttributionView } from '../../state/selectors/attribution-view-resource-selectors';
-import {
-  getAttributionIdOfDisplayedPackageInManualPanel,
-  getSelectedResourceId,
-} from '../../state/selectors/audit-view-resource-selectors';
+import { getSelectedAttributionIdInAttributionView } from '../../state/selectors/all-views-resource-selectors';
+import { getAttributionIdOfDisplayedPackageInManualPanel } from '../../state/selectors/all-views-resource-selectors';
+import { getSelectedResourceId } from '../../state/selectors/audit-view-resource-selectors';
 import {
   getPopupAttributionId,
   getSelectedView,
