+++ src/Frontend/state/selectors/attribution-view-resource-selectors.ts
@@ -3,13 +3,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 import { State } from '../../types/types';
+import { getSelectedAttributionIdInAttributionView } from './all-views-resource-selectors';
 import { getManualAttributionsToResources } from './all-views-resource-selectors';
-
-export function getSelectedAttributionIdInAttributionView(
-  state: State,
-): string {
-  return state.resourceState.attributionView.selectedAttributionId;
-}
 
 export function getTargetSelectedAttributionId(state: State): string | null {
   return state.resourceState.attributionView.targetSelectedAttributionId;
